1. updated to compile on windows with VC9.  In particular, made c89 compatabile and added a config.w32 for normal windows extension builds. 2. Fixed 2 very significant memory leaks during video decoding.  3.  Fixed serious frame decode issue for B frames where random end of video frames would get skipped (at least in current ffmpeg implementations see https://github.com/chelyaev/ffmpeg-tutorial/issues/7 and thanks to sobotka for help).  4.  Tried to fix fps and frame count calcuations though don't know if they work across codecs.

+++ ffmpeg_frame.h
@@ -65,7 +65,7 @@ typedef struct {
     int height;
     int pixel_format;
     int keyframe;
-    int64_t pts;
+    double pts;
 } ff_frame_context;
 
 ff_frame_context* _php_create_ffmpeg_frame(INTERNAL_FUNCTION_PARAMETERS);
